% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airpollute}
\alias{airpollute}
\title{Air Pollution Dataset}
\format{
\subsection{\code{airpollute}}{

A data frame with 8760 rows and 5 columns:
}
}
\source{
<Dahlhaus and Eichler (2003)>
}
\usage{
data(airpollute)
}
\description{
Time series dataset from Dahlhaus and Eichler (2003) and
Davis et al. (2016). Recorded variables include \code{CO} and \code{NO}
(pollutants mainly emitted from the cars and industry),
\code{NO} and \code{O} (generated from different reactions in the atmosphere),
and the global solar radiation intensity \code{R}.
}
\keyword{datasets}
