% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMCR.R
\name{aMCR}
\alias{aMCR}
\title{Implements sparse multivariate regression}
\usage{
aMCR(
  Y,
  X,
  lam.vec = NULL,
  nfolds = 10,
  nlambda = 100,
  penalty = c("lasso", "MCP", "SCAD"),
  symm.rule = TRUE,
  seed = 123
)
}
\arguments{
\item{Y}{multivariate response}

\item{X}{data matrix}

\item{lam.vec}{vector of lambda values. None by default.}

\item{nfolds}{number of folds; 10 by default}

\item{nlambda}{number of lambdas; 100 by default}

\item{penalty}{penalty type; lasso, MCP and SCAF. lasso is the default}

\item{symm.rule}{the rule to make Gamma_hat matrix symmetric, if True \code{or} rule is used, and \code{and} rule otherwise}

\item{seed}{random seed}
}
\value{
*\code{B_hat} - Estimated coefficient matrix
*\code{Gamma_hat} - Estimated gamma matrix with the same sparsity as the precision matrix
*\code{B_0_hat} - Estimated initial coefficient matrix
*\code{lambda.min} - the vector of lambda values selected from cross-validation
}
\description{
This function implements joint estimation of sparse multivariate
regression and conditional graphical models as in Wang (2015)
(https://www3.stat.sinica.edu.tw/statistica/oldpdf/A25n31.pdf)
}
