% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{back_code}
\alias{back_code}
\title{Back code OTHERS into pre-coded question}
\usage{
back_code(raw, coded)
}
\arguments{
\item{raw}{dataframe with 2 columns}

\item{coded}{dataframe with 2 columns}
}
\description{
This function is to back code verbatim into a pre-coded question
}
\details{
The raw and coded dataframes should have one common column for SN matching. If no need for SN matching, just populate both SN columns with the same vector.
The function works by replacing code "97" in the raw data with codes in the coded data, given that the manually coded column is not empty.
}
\examples{
raw <- data.frame(SN=c(1, 2000, 3, 4), raw_data=c("1,2,97", "1,3", "97", "1,2,97"), stringsAsFactors=FALSE)##Populate a dataframe
coded <- data.frame(SN=c(2000, 1, 3, 4), coded_data=c(NA, "2,3", "9", "97"), stringsAsFactors=FALSE)##Populate another dataframe
##You will see the output that shows 1) all duplicate should be removed and 2) SA question is also applicable
back_code(raw, coded)

}
