% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section1_biasplot.R
\name{bias.plot}
\alias{bias.plot}
\title{Plots the bias for Lasso, and MCP as described in Chapter 1
Section 4.2}
\usage{
bias.plot(n.list, beta, nsim = 100, seed = 1234)
}
\arguments{
\item{n.list}{list of n values}

\item{beta}{coefficient matrix}

\item{nsim}{number of simulations}

\item{seed}{random seed}
}
\value{
*\code{bias.lasso} - estimated bias for Lasso
*\code{bias.mcp}     - estimated bias for MCP
}
\description{
Plots the bias for Lasso, and MCP as described in Chapter 1
Section 4.2
}
