% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brand}
\alias{brand}
\title{Car brand perception survey.}
\format{
A data frame with 1000 rows and 16 variables.
}
\usage{
data(brand)
}
\description{
The data is based on a survey of customers
from Nissan's target segments, described as people between 25 and 35 with annual household incomes between
50000 and 100000. Participants were asked about their perceptions (from scale 1 to 7) of 10 different cars
(Infinity G20, Ford T-bird, Audi 90, Toyota Supra, Eagle Talon, Honda Prelude, Saab 900, Pontiac Firebird,
BMW 318i, Mercury Capri) on 15 attributes
}
\keyword{datasets}
