% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardiodata}
\alias{cardiodata}
\title{Dataset from cardiovascular microarray study}
\format{
\subsection{\code{cardiodata}}{

A data frame with 20436 rows and 63 columns:
}
}
\source{
<Efron. B. Large-Scale InferenceL Empirical Bayes
Methods for Estimation, Testing, and Prediction (2010)>
}
\usage{
cardiodata
}
\description{
The dataset has 63 subjects with 44 healthy controls, and
19 cardiovascular patients, and 20436 genes measured for
each subject.
}
\keyword{datasets}
