% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{check_unique}
\alias{check_unique}
\title{Check if the variable has unique answer}
\usage{
check_unique(vector, unique_code)
}
\arguments{
\item{vector}{variable to be checked}

\item{unique_code}{specify unique code}
}
\value{
vector
}
\description{
This function is to check if the variable contains unique answer. Return TRUE when it does while return FALSE when it does not. For example, DK (code 98) should exist in a MA question alone; if DK (code 98) exist along with others codes in a MA question, the function will return FALSE.
}
\examples{
check_unique(c("1,2,3", "1,2", "1"), 1)#FALSE FALSE TRUE

}
