% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemo}
\alias{chemo}
\title{Chemometrics example from Skagerberg et al. (1992).}
\format{
A list with two elements: \code{X} frame with 56 rows and 22 variables, and
\code{y} frame with 56 rows and 6 variables.
}
\source{
Skagerberg et al. (1992)
}
\usage{
data(chemo)
}
\description{
consists of \code{n = 56} observations, each with \code{p = 22} predictor variables and
\code{q = 6} responses. The data are taken from a simulation of a low density tubular polyethylene
reactor The predictor variables consist of temperatures measured at equal distances along the
reactor together with the wall temperature of the reactor and the feed rate.
The responses are the output characteristics of the produced polymers
}
\keyword{datasets}
