% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{combine}
\alias{combine}
\title{Combine 2 questions together}
\usage{
combine(vector1, vector2)
}
\arguments{
\item{vector1}{character vector}

\item{vector2}{character vector}
}
\description{
This function is to combine 2 questions together. The returned value should have no duplicate and excessive comma.
}
\examples{
vector1 <- c("1", "2", NA, NA)
vector2 <- c("3", "2,3,4", NA, "1")
combine(vector1, vector2) #"1,3"   "2,3,4" ""      "1"
}
