% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparegraph.R
\name{comparegraph}
\alias{comparegraph}
\title{Compare graphs
Part of this function is borrowed from \code{pcalg} package.}
\usage{
comparegraph(est, true, verbose = FALSE)
}
\arguments{
\item{est}{estimated weighted adjacency matrix}

\item{true}{true weighted adjacency matrix}

\item{verbose}{display the result; default is false}
}
\value{
f1, tpr, fpr, and tdr
}
\description{
Compare graphs
Part of this function is borrowed from \code{pcalg} package.
}
