% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covshrink.R
\name{cov_shrink_F}
\alias{cov_shrink_F}
\title{Implements shrinkage estimation of the covariance matrix using case F as in
Schaffer and Strimmer (2005)}
\usage{
cov_shrink_F(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
*\code{Sigma} - shrinkage covariance matrix
}
\description{
This function implements shrinkave covariance matrix estimation
using case F as in Schaffer and Strimmer (2005) (https://strimmerlab.github.io/publications/journals/shrinkcov2005.pdf)
}
