% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covEstimation.R
\name{covEstimation}
\alias{covEstimation}
\title{Implement covariance matrix estimation}
\usage{
covEstimation(
  X,
  k,
  method = c("thresholding", "banding", "tapering"),
  thr.method = c("adaptive", "soft", "hard"),
  demean = FALSE,
  scale = FALSE
)
}
\arguments{
\item{X}{data matrix}

\item{k}{tuning parameter}

\item{method}{if \verb{method = "thresholding} and the \code{thr.method} is \code{adaptive} implements
method by Cai and Liu (2011), otherwise method by Bickel and Levina (2008b).}

\item{thr.method}{method for thresholding}

\item{demean}{demeanize the data. True by default}

\item{scale}{standardize the data. False by defaults.}
}
\value{
*\code{Sigma} - Estimated covariance matrix
*\code{S}     - Sample covariance matrix
}
\description{
This function implements covariance matrix estimation
using:
\itemize{
\item Thresholding method proposed in Cai and Liu (2011) and Bickel and Levina (2008b)
\item Banding method proposed in Bickel and Levina (2008a)
\item Tapering method proposed in Cai et al (2010)
}
}
