% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{Information related to the COVID-19 infections for each
date-region pair}
\format{
\subsection{\code{covid}}{

Covid data
}
}
\source{
\url{https://github.com/GoogleCloudPlatform/covid-19-open-data/blob/main/docs/table-epidemiology.md}
}
\usage{
covid
}
\description{
Information related to the COVID-19 infections for each
date-region pair
}
\keyword{datasets}
