% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covEstimation.R
\name{cv.covEstimation}
\alias{cv.covEstimation}
\title{Implement covariance matrix estimation}
\usage{
cv.covEstimation(
  X,
  k.list = NULL,
  n.k = 30,
  k.min = 0.01,
  nfolds = 5,
  method = c("thresholding", "banding", "tapering"),
  thr.method = c("adaptive", "soft", "hard"),
  demean = FALSE,
  scale = FALSE,
  k.max = NULL
)
}
\arguments{
\item{X}{data matrix}

\item{k.list}{list of tuning parameter; default is NULL}

\item{n.k}{number of tuning parameters; default is 30}

\item{k.min}{minimum value for tuning parameter; default is 0.01}

\item{nfolds}{number of folds for CV; default is 5}

\item{method}{if \verb{method = "thresholding} and the \code{thr.method} is \code{adaptive} implements
method by Cai and Liu (2011), otherwise method by Bickel and Levina (2008b).}

\item{thr.method}{method for thresholding}

\item{demean}{demeanize the data. True by default}

\item{scale}{standardize the data. False by defaults.}

\item{k.max}{maximum value for tuning parameter; no maximum is specified by default}
}
\value{
*\code{err_fit} - fitted errors
*\verb{cv,k} - selected tuning parameter
*\code{Sigma_fit} - Estimated covariance matrix
*\code{S}     - Sample covariance matrix
}
\description{
This function implements covariance matrix estimation
using:
\itemize{
\item Thresholding method proposed in Cai and Liu (2011) and Bickel and Levina (2008b)
\item Banding method proposed in Bickel and Levina (2008a)
\item Tapering method proposed in Cai et al (2010)
}
}
