% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshglasso.R
\name{cv.threshlasso}
\alias{cv.threshlasso}
\title{Implement and select threshold for thresholded Lasso}
\usage{
cv.threshlasso(
  x,
  y,
  lambda = NULL,
  ngrid = 30,
  min.thresh = NULL,
  max.thresh = NULL,
  grid = NULL,
  nfold = 5,
  seed = 1234,
  fold_ids = NULL,
  ...
)
}
\arguments{
\item{x}{data matrix}

\item{y}{dependent variable}

\item{lambda}{penalization parameter; default is 0.01}

\item{ngrid}{number of grids for tuning threshold; default is 30}

\item{min.thresh}{minimum threshold value; default is 0.01}

\item{max.thresh}{maximum threshold value; default is maximum non-diagonal value of sample covariance matrix}

\item{grid}{user specified grid}

\item{nfold}{number of folds for CV}

\item{seed}{random seed}

\item{fold_ids}{(optional) vector of length n specifiying the folds assignment (from 1 to max(folds_ids)), if supplied the value of k is ignored}

\item{...}{other glmnet options}
}
\value{
coefficient matrix
}
\description{
This function implements tuning of thresholded Lasso
using cross-validation
}
