% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{dup_remove_no_message}
\alias{dup_remove_no_message}
\title{Remove duplicate(s)}
\usage{
dup_remove_no_message(vector)
}
\arguments{
\item{vector}{character vector}
}
\description{
This function is to remove any duplicate in MA question
}
\examples{
dup_remove(c("1,2,3,2,3,3"))

}
