% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{excess_comma}
\alias{excess_comma}
\title{Remove excessive comma}
\usage{
excess_comma(vector)
}
\arguments{
\item{vector}{character vector with excessive comma}
}
\value{
character vector without excessive comma
}
\description{
This function is to remove any trailing, leading as well as consecutive comma in between.
}
\examples{
excess_comma(c(",,1,,,2,,3,,,"))

}
