% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fat}
\alias{fat}
\title{Percentage of Body Fat and Body Measurements}
\format{
\subsection{\code{fat}}{

A data frame with 252 observations on the following 18 variables:
}
}
\source{
<Johnson R. Journal of Statistics Education v.4, n.1 (1996)>
}
\usage{
fat
}
\description{
Age, weight, height, and 10 body circumference measurements
are recorded for 252 men. Each man's percentage of body fat was
accurately estimated by an underwater weighing technique.
}
\keyword{datasets}
