% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covEstimation.R
\name{find.thr.min}
\alias{find.thr.min}
\title{Finds the threshold parameter \code{k.min} that makes the thresholded covariance
matrix positive definite}
\usage{
find.thr.min(
  X,
  k.grid,
  thr.method = c("adaptive", "soft", "hard"),
  demean = FALSE,
  scale = FALSE
)
}
\arguments{
\item{X}{data matrix}

\item{k.grid}{grid of tuning parameter}

\item{thr.method}{method for thresholding}

\item{demean}{demeanize the data. True by default}

\item{scale}{standardize the data. False by defaults.}
}
\value{
*\code{k.min} - selected threshold
*\code{eigen.store}     - stored minimum eigenvalues for each k in the \code{k.grid}
*\code{k.grid}          - grid of tuning parameters
}
\description{
Finds the threshold parameter \code{k.min} that makes the thresholded covariance
matrix positive definite
}
