% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodrisk}
\alias{floodrisk}
\title{Fictional flood risk dataset}
\format{
A data with 500 rows and 2 variables:
\describe{
\item{\code{rainfall (mm)}}{double the amounf of rain fall}
\item{\code{riveflow (m3/s)}}{double river flow rate}
}
}
\usage{
floodrisk
}
\description{
Fictinal flood risk dataset with two variables
rainfall and river flow. The dataset is generated using the following:
\code{set.seed(123)}
\code{n <- 500}
\code{rainfall <- rgamma(n, shape = 5, rate = 0.2)}
\code{epsilon <- rnorm(n, mean = 0, sd = 0.2)}
\code{riverflow <- exp(0.05 * rainfall + epsilon)}
}
\keyword{datasets}
