% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genclique.R
\name{genclique}
\alias{genclique}
\title{Generate a block diagonal covariance matrix
This function is borrowed from \code{spcov} package.
For details, see https://cran.r-project.org/web/packages/spcov/index.html.}
\usage{
genclique(ncliques = 5, cliquesize = 5, theta = 1, seed = 1234)
}
\arguments{
\item{ncliques}{number of blocks; default is 5}

\item{cliquesize}{size of each block; default is 5}

\item{theta}{magnitude of non-zeros elements; default is 1}

\item{seed}{random seed}
}
\value{
\itemize{
\item \code{Sigma} p x p covariance matrix
\item \code{A} p x p square root of  \code{Sigma}
}
}
\description{
Generate a block diagonal covariance matrix
This function is borrowed from \code{spcov} package.
For details, see https://cran.r-project.org/web/packages/spcov/index.html.
}
