% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genclique.R
\name{gendataclique}
\alias{gendataclique}
\title{Generate data from a block diagonal covariance matrix}
\usage{
gendataclique(p = 20, n = 100, ncliques = 5, theta = 1, seed = 1234)
}
\arguments{
\item{p}{number of features}

\item{n}{number of observations}

\item{ncliques}{number of blocks; default is 5}

\item{theta}{magnitude of non-zeros elements; default is 1}

\item{seed}{random seed}
}
\value{
\itemize{
\item \code{X} n x p data matrix
\item \code{S} p x p sample covariance matrix
\item \code{Sigma} p x p true covariance matrix
\item \code{A} p x p square root of  \code{Sigma}
}
}
\description{
Generate data from a block diagonal covariance matrix
}
