% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section3_fstat.R
\name{generate_cardio_fstat}
\alias{generate_cardio_fstat}
\title{Generates matrix of F-statistics for cardiodata example in Chapter 3
Section 2}
\usage{
generate_cardio_fstat(data, top = 50, last = 100)
}
\arguments{
\item{data}{cardiodata}

\item{top}{select the highest \code{top} number of F-stat; default is 50}

\item{last}{select the lowest \code{last} number of F-stat; default is 100}
}
\value{
*\code{df} generated F-stat
}
\description{
Generates matrix of F-statistics for cardiodata example in Chapter 3
Section 2
}
