% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{generatedata}
\alias{generatedata}
\title{Generate data from AD or Star shape undirected graph}
\usage{
generatedata(
  n,
  p,
  type = c("star", "AD1"),
  variance = 1,
  rho = 0.5,
  AD_order = 1,
  seed = 123
)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{type}{type of the undirected graph}

\item{variance}{variance of the generated data; default is 1}

\item{rho}{defines edge weights}

\item{AD_order}{order of the AD process; ignored for \code{type(star)}}

\item{seed}{random seed}
}
\value{
*\code{omega} p x p precision matrix
\itemize{
\item \code{data} n x p data
}
}
\description{
Generate data from AD or Star shape undirected graph
}
