% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genlatentdata}
\alias{genlatentdata}
\title{Generate data from latent variable graphical model}
\usage{
genlatentdata(
  p_o = 10,
  p_l = 5,
  N = 150,
  prob_o = 0.4,
  prob_l = 0.2,
  prob_ol = 0.7,
  lwb = 0.4,
  seed = 1234
)
}
\arguments{
\item{p_o}{number of observed variables; default is 10}

\item{p_l}{number of latent variables; default is 5}

\item{N}{number of observations; default is 150}

\item{prob_o}{probability two observed variables are connected; default is 0.4}

\item{prob_l}{probability two latent variables are connected; default is 0.2}

\item{prob_ol}{probability the latent and observed variables are connected; default is 0.2}

\item{lwb}{lower value for the elements of the inverse covariance matrix}

\item{seed}{random seed}
}
\value{
\itemize{
\item \code{X} N x p_o data matrix
\item \code{theta_o} true observed inverse covariance matrix
\item \code{theta_obs} observed inverse covariance matrix
}
}
\description{
Generate data from latent variable graphical model
}
