% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{get_tab_var}
\alias{get_tab_var}
\title{Get variables for tabulation}
\usage{
get_tab_var(start_with)
}
\arguments{
\item{character}{character}
}
\value{
list
}
\description{
This function is to get variables for tabulation
}
\examples{
get_tab_var("Q")

}
