% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalbank}
\alias{globalbank}
\title{The global bank stock return volatility data}
\format{
\subsection{\code{globalbank}}{

A data frame with 2672 rows and 54 columns:
}
}
\source{
<Demiret et. al. (2018)>
}
\usage{
data(globalbank)
}
\description{
A subset of data from the Demirer, M., F. X. Diebold, L. Liu, and K. Yilmaz. 2018. Estimating global bank
network connectedness. Journal of Applied Econometrics 33(1).
Original data contains 96 banks from 29 developed and emerging economies (countries) from
September 12, 2003, to February 7, 2014. The subset contains only
economies where the number of banks in each economy is greater than 4, total of 54
banks.
Below we provide details on bank names and the country they belong:
Variable       Bank        Country
hsba.ln        HSBC        Great Britain (gb)
X8306.to       MTBH        Japan (jp)
jpm            JPM         USA (us)
barc.ln        BARC        Great Britain (gb)
bac            BAC         USA (us)
c              Citi        USA (us)
X8411.to       MZH         Japan (jp)
rbs.ln         RBS         Great Britain (gb)
X8316.to       SMTM        Japan (jp)
wfc            WFC         USA (us)
lloy.ln        LLOY        Great Britain (gb)
ucg.mi         UCG         Italy (it)
gs             GS          USA (us)
isp.mi         ISP         Italy (it)
MS             MS          USA (us)
td.t           TD          Canada (ca)
ry.t           RBC         Canada (ca)
nab.au         NAB         Australia (au)
bns.t          BNS         Canada (ca)
cba.au         CBA         Australia (au)
stan.ln        STAN        Great Britain (gb)
X600036.sh     CMB         China (cn)
anz.au         ANZ         Australia (au)
wbc.au         WBC         Australia (au)
X600000.sh     SHGO        China (cn)
X600016.sh     CMB         China (cn)
bmo.t          BMO         Canada (ca)
X8308.to       RSNH        Japan (jp)
X8604.to       NMRH        Japan (jp)
X8309.to       SMTM        Japan (jp)
cm.t           CBC         Canada (ca)
bk.us          BNYM        USA (us)
usb            USB         USA (us)
pnc.us         PNC         USA (us)
X000001.sz     PAB         China (cn)
cof            COF         USA (us)
X600015.sh     HXB         China (cn)
bmps.mi        BMPS        Italy (it)
stt.us         SSC         USA (us)
BBT            BBT         USA (us)
NA.T           NBC         Canada (ca)
STI.US         STB         USA (us)
bp.mi          BP          Italy (it)
mqg.au         MQG         Australia (au)
X8354.to       FFG         Japan (jp)
X8332.to       BOY         Japan (jp)
fitb.us        FTB         USA (us)
rf.us          RF          USA (us)
X8331.to       CBB         Japan (jp)
uni.mi         UGF         Italy (it)
X8377.to       HFG         Japan (jp)
X8355.to       ShB         Japan (jp)
mb.mi          MBCF        Italy (it)
X8418.to       YFG         Japan (jp)
}
\keyword{datasets}
