% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.even.R
\name{is.even}
\alias{is.even}
\title{Check if numbers are even}
\usage{
is.even(x = NULL)
}
\description{
Check if numbers are even
}
\examples{
is.even(2)
is.even(5)
is.even(-4)
is.even(0)

# use vectors
is.even(  c(5,3,-2,0,4) )

}

