% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.odd.R
\name{is.odd}
\alias{is.odd}
\title{Check if numbers are odd}
\usage{
is.odd(x = NULL)
}
\description{
Check if numbers are odd
}
\examples{
is.odd(2)
is.odd(5)
is.odd(-4)
is.odd(0)

# use vectors
is.odd(  c(5,3,-2,0,4) )

}

