% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if numbers of integer value (not type)}
\usage{
is.wholenumber(x = NULL)
}
\arguments{
\item{x}{numeric object}
}
\description{
Check if numbers of integer value (not type)
}
\examples{
is.wholenumber(2)
is.wholenumber(0.9)
is.wholenumber(5)
is.wholenumber(-3)
is.wholenumber(0)

# use vectors
is.wholenumber( c(5,1.3,-2,0,-3.7,4) )

}

