% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{mentions}
\alias{mentions}
\title{Get the number of mentions in a MA response}
\usage{
mentions(vector)
}
\arguments{
\item{vector}{character vector (MA question)}
}
\description{
This function is to Get the number of mentions in a MA response
}
\examples{
mentions(c("1", "2,4", "3,5"))#1 2 2

}
