% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oslo2rppa}
\alias{oslo2rppa}
\title{Oslo 2 data}
\format{
\subsection{\code{oslo2rppa}}{

A data frame with 280 rows and 100 columns:
}
}
\source{
<Lingjarde, C., Lien, T.G., Borgan, O. et al.
Tailored graphical lasso for data integration in gene network reconstruction.
BMC Bioinformatics 22, 498 (2021). https://doi.org/10.1186/s12859-021-04413-z>
}
\usage{
data(oslo2rppa)
}
\description{
Includes n = 280 breast cancer samples collected from hospitals in Oslo
The gene expression for p=100 genes has been quantified by measuring their proteins using RPPA technology.
}
\keyword{datasets}
