% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covEstimation.R
\name{plot_mat}
\alias{plot_mat}
\title{Plot the covariance matrix. Borrowed from \code{varband} package.}
\usage{
plot_mat(Mat, main = NULL)
}
\arguments{
\item{Mat}{covariance matrix}

\item{main}{title of the plot}
}
\description{
Plot the covariance matrix. Borrowed from \code{varband} package.
}
