% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshglasso.R
\name{predict_threshlasso}
\alias{predict_threshlasso}
\title{Predicts thresholded Lasso}
\usage{
predict_threshlasso(thresh, test_x)
}
\arguments{
\item{thresh}{object from the \code{cv.threshlasso}}

\item{test_x}{new dataset for prediction}
}
\value{
vector of predictions
}
\description{
Predicts thresholded Lasso
}
