% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psych}
\alias{psych}
\title{Hypothetical datasets.}
\format{
\code{X} frame with 600 rows and 6 variables.
}
\source{
Skagerberg et al. (1992)
}
\usage{
data(psych)
}
\description{
hypothetical dataset with 600 observations on seven variables.
The psychological variables are locus of control (\code{locus_of_control}),
self-concept (\code{self_concept}), and motivation (\code{motivation}). The academic variables
are standardized tests scores in reading (\code{read}), writing (\code{write}),
and science (\code{science}).
}
\keyword{datasets}
