% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{rank_trans}
\alias{rank_trans}
\title{Transform ranking variables}
\usage{
rank_trans(vector, no_R)
}
\arguments{
\item{vector}{character vector}

\item{no_R}{single integer}
}
\description{
This function is to transform ranking variables to comma-separated MA question
}
\details{
Argument "no_R" indicates how many attributes the respondents should rank
When codes of attributes are not consecutive, e.g. attribute 4 corresponds to code 8, instead of code 4, function replace_code() can be used to replace code 4 by code 8
}
\examples{
#Create a ranking dataframe
ranking_df = data.frame(attribute1 = c("3", "3", "2", "2"), attribute2 = c("1", "1", "1", "3"), attribute3 = c("2", "2", "4", "1"), attribute4 = c("4", "4", "3", "4"))
#Apply the function rowwise
apply(ranking_df, 1, rank_trans, no_R = 4)#"2,3,1,4" "2,3,1,4" "2,1,4,3" "3,1,2,4"; the first response means rank 1, the second means rank 2 and so on

}
