% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{remove_code}
\alias{remove_code}
\title{Recodes or transforms data collected from market reserach study}
\usage{
remove_code(vector, remove)
}
\arguments{
\item{vector}{}

\item{remove}{}
}
\description{
This package performs various operations in data transfomation commonly seen in market reserach study
}
\examples{

}
