% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{replace_code}
\alias{replace_code}
\title{Replace particular code(s)}
\usage{
replace_code(vector, to_be_replaced, replacement)
}
\arguments{
\item{vector}{character vector (MA question) or single numeric value (SA question)}

\item{to_be_replaced}{character string (only 1 code to be replaced) or character vector (more than 1 code to be replaced)}

\item{replacement}{character string (only 1 code as replacement) or character vector (more than 1 code as replacement)}
}
\description{
This function is to replace code(s) in either SA or MA question
}
\examples{
replace_code(c("1", "2,4", "3,5"), 1, 2)#"2"   "2,4" "3,5"
replace_code(c(1, 2, 3), 1, 99)#99  2  3

}
