% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{reverse_code}
\alias{reverse_code}
\title{Reverse codes}
\usage{
reverse_code(vector)
}
\arguments{
\item{vector}{character vector}
}
\description{
This function is to reverse codes in SA question
}
\details{
Note that for the sake of subsequent processing of data, the output of this function is numeric
}
\examples{
a = c("1", "2", "3", "4", "5", "9", NA)
reverse_code(a) #9  8  7  6  5  1 NA

}
