% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sachs}
\alias{sachs}
\title{Flow cytometry data from Sachs et al. (2005).}
\format{
A data frame with 7466 rows and 11 variables:
}
\source{
Sachs et al. (2005)
}
\usage{
data(sachs)
}
\description{
A dataset consists of i $n = 7466$ simultaneous measurements of $p = 11$
phosphorylated proteins and phospholipids.
}
\keyword{datasets}
