\name{RECODE-package}
\alias{RECODE-package}
\alias{RECODE}
\docType{package}
\title{
\packageTitle{RECODE}
}
\description{
\packageDescription{RECODE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RECODE}
\packageIndices{RECODE}
}
\author{
\packageAuthor{RECODE}

Maintainer: \packageMaintainer{RECODE}
}
\keyword{scRNE-seq}
\seealso{
}