% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{shifter}
\alias{shifter}
\title{Shift elements in vector by specific distance and direction}
\usage{
shifter(vector = x, distance = n, direction = c("left", "right"))
}
\arguments{
\item{vector}{numeric or character vector}

\item{distance}{integer}

\item{direction}{c("left", "right")}
}
\description{
This function is to shift elements in vector by specific distance and direction
}
\examples{
shifter(1:4, 0, "left")#1 2 3 4
shifter(1:4, 1, "left")#2 3 4 1
shifter(1:4, 1, "left")#Distance to be shifted exceeds length of vector

}
