% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Log returns of 8 stocks}
\format{
\subsection{\code{stocks}}{

A data frame with 314 rows and 8 columns:
}
}
\source{
YahooFinance
}
\usage{
data(stocks)
}
\description{
Log-returns of eight US stocks (Walmart, Exxon, Ford. GE, Conoco Phillips,Citigroup,
IBM, and AIG) from 2017 to 2022. The dataset consists of 314 observations
}
\keyword{datasets}
