% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{sum_t}
\alias{sum_t}
\title{Fabricate header and stub for the generation of Summary Table}
\usage{
sum_t(dataframe)
}
\arguments{
\item{dataframe}{dataframe}
}
\description{
This function is to fabricate header and stub for the generation of Summary Table
}
\details{
Use in combination with the following spec in SC:

Q30M             Header_Qno. Label

C1               Att1

C2               Att2

C3               Att3


Q31C   Qno*      Stub_Qno. Label

*Qno where individual att come up
}
\examples{

att1 = c ("1", "2", "3", "4", "5")
att2 = c (NA, "2", "3", NA, "5")
att3 = c (NA, NA, NA, NA, NA)
att4 = c ("1", "2", "4", "3", "1")
att5 = c ("1", "2", "5", NA, "5")

df = data.frame(att1, att2, att3, att4, att5)

sum_t(df)

}
