% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SW2002}
\alias{SW2002}
\title{Stock and Watson (2002) dataset.}
\format{
A data frame with 215 rows and 146 variables.
}
\source{
http://www.princeton.edu/~mwatson/ddisk/wc00.zip
}
\usage{
data(SW2002)
}
\description{
A dataset containing 215 monthly US macroeconomic variables spanning the period 1959-1998.
The dataset is borrowed from  Stock and Watson (2002).
}
\keyword{datasets}
