% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshglasso.R
\name{threshglasso}
\alias{threshglasso}
\title{Implement and select threshold for thresholded GLasso}
\usage{
threshglasso(
  X = NULL,
  lambda = 0.01,
  ngrid = 30,
  min.thresh = 0.01,
  max.thresh = NULL,
  grid = NULL,
  tol = 1e-05,
  gamma = 0.5,
  diagonal = FALSE,
  crit = c("BIC", "AIC", "eBIC"),
  maxit = 1000,
  start = c("warm", "cold"),
  ...
)
}
\arguments{
\item{X}{data matrix}

\item{lambda}{penalization parameter; default is 0.01}

\item{ngrid}{number of grids for tuning threshold; default is 30}

\item{min.thresh}{minimum threshold value; default is 0.01}

\item{max.thresh}{maximum threshold value; default is maximum non-diagonal value of sample covariance matrix}

\item{grid}{user specified grid}

\item{tol}{tolerance for GLasso}

\item{gamma}{parameter for eBIC; default is 0.5}

\item{diagonal}{pernalize diagonal}

\item{crit}{criteria for threshold selection}

\item{maxit}{number of iterations}

\item{start}{\code{warm} or \code{cold} start to improve convergence}

\item{...}{other options}
}
\value{
estimated (inverse) covariance matrix
}
\description{
This function implements tuning of thresholded GLasso
using eBIC, AIC, and BIC
}
