% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yearpredictionmsd}
\alias{yearpredictionmsd}
\title{Part of Million Song dataset.}
\format{
A data frame with 10000 rows and 91 variables.
}
\source{
https://archive.ics.uci.edu/dataset/203/yearpredictionmsd
}
\usage{
data(yearpredictionmsd)
}
\description{
A dataset consists of information about 10000 songs, which are mostly western and range from
1922 to 2011. It contains p = 90 timbre features.
The dataset is borrowed from  https://archive.ics.uci.edu/dataset/203/yearpredictionmsd.
}
\keyword{datasets}
