% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{back_code}
\alias{back_code}
\title{Recodes or transforms data collected from market reserach study}
\usage{
back_code(raw, coded)
}
\arguments{
\item{raw}{}

\item{coded}{}
}
\value{
recoded
}
\description{
This package performs various operations in data transfomation commonly seen in market reserach study
}
\examples{

}
