% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{back_code}
\alias{back_code}
\title{Back code OTHERS into pre-coded question}
\usage{
back_code(raw, coded, others_code, SN_matching)
}
\arguments{
\item{raw}{dataframe with 2 columns (for backcoding with SN matching); vector (for backcoding without SN mathcing)}

\item{coded}{dataframe with 2 columns (for backcoding with SN matching); vector (for backcoding without SN mathcing)}

\item{others_code}{Single integer}

\item{SN_matching}{Logical}
}
\description{
This function is to back code verbatim into a pre-coded question
}
\details{
Inputs of this function vary upon usage; If no need for SN matching, set FALSE in the argument "SN_matching", or vice
}
\examples{
raw <- data.frame(SN=c(1, 2000, 3, 4), raw_data=c("1,2,97", "1,3", "97", "1,2,97"), stringsAsFactors=FALSE)##Populate a dataframe
coded <- data.frame(SN=c(2000, 1, 3, 4), coded_data=c(NA, "2,3", "9", "97"), stringsAsFactors=FALSE)##Populate another dataframe
##You will see the output that shows 1) all duplicate should be removed and 2) SA question is also applicable
back_code(raw, coded)

}
