% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{dup_remove}
\alias{dup_remove}
\title{Recodes or transforms data collected from market reserach study}
\usage{
dup_remove(vector)
}
\arguments{
\item{vector}{}
}
\value{
results
}
\description{
This package performs various operations in data transfomation commonly seen in market reserach study
}
\examples{

}
