% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{excess_comma}
\alias{excess_comma}
\title{Recodes or transforms data collected from market reserach study}
\usage{
excess_comma(vector)
}
\arguments{
\item{vector}{gggg}
}
\value{
results
}
\description{
This package performs various operations in data transfomation commonly seen in market reserach study
}
\details{
ggg
}
\examples{

}
