% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{remove_code}
\alias{remove_code}
\title{Remove particular code(s)}
\usage{
remove_code(vector, remove)
}
\arguments{
\item{vector}{character vector (MA question) or single numeric value (SA question)}

\item{remove}{character string (only 1 code to remove) or character vector (more than 1 code to remove)}
}
\description{
This function is to remove code(s) in either SA or MA question
}
\examples{
remove_code(c("1", "2,4", "3,5"), remove=c(2,3))#remove more than 1 code in MA question; "1" "4" "5"
remove_code(c("1", "2,4", "3,5"), remove=2)#remove only 1 code in MA question; "1"   "4"   "3,5"
remove_code(c(1, 2, 3), remove=3)#remove 1 code in SA question; "1" "2" ""

}
