% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{remove_code_no_message}
\alias{remove_code_no_message}
\title{Remove particular code(s)}
\usage{
remove_code_no_message(vector, remove)
}
\arguments{
\item{vector}{}

\item{remove}{}
}
\description{
This function is to remove code(s) in either SA or MA question
}
\examples{

}
