% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeR.R
\name{rotate_JS}
\alias{rotate_JS}
\title{Create rotational combination in JS array format}
\usage{
rotate_JS(vector = x, direction = c("left", "right"), keep_zero)
}
\arguments{
\item{vector}{numeric vector}

\item{direction}{c("left", "right")}

\item{keep_zero}{logical}
}
\description{
This function is exclusively used for Rail Gen 2.0 project series to create rotational combination in JS array format
}
\examples{
rotate_JS(c(1:4), "right", keep_zero=TRUE)#"1:[0,1,2,3,4]," "2:[0,4,1,2,3]," "3:[0,3,4,1,2]," "4:[0,2,3,4,1],"
rotate_JS(c(1:4), "left", keep_zero=TRUE)#"1:[0,1,2,3,4]," "2:[0,2,3,4,1]," "3:[0,3,4,1,2]," "4:[0,4,1,2,3],"

}
