\name{scRECODE-package}
\alias{scRECODE-package}
\alias{scRECODE}
\docType{package}
\title{
\packageTitle{scRECODE}
}
\description{
\packageDescription{scRECODE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scRECODE}
\packageIndices{scRECODE}
}
\author{
\packageAuthor{scRECODE}

Maintainer: \packageMaintainer{scRECODE}
}
\keyword{scRNE-seq}
\seealso{
}